Move into new file and add panic unwind tests
